restructure(Args): change InputTypes to ArgsTypes and remove unnecessary types and decorators
