+++ demo/api/schema.gql
@@ -379,7 +379,7 @@ type Query {
   redirect(id: ID!): Redirect!
   redirectSourceAvailable(scope: RedirectScopeInput!, source: String!): Boolean!
   damItemsList(offset: Int = 0, limit: Int = 20, sortColumnName: String, sortDirection: SortDirection = ASC, folderId: ID, includeArchived: Boolean, filter: DamItemFilterInput): PaginatedDamItems!
-  damItemListPosition(id: ID!, type: DamItemType!, args: DamItemPositionInput!): Float!
+  damItemListPosition(sortColumnName: String, sortDirection: SortDirection = ASC, id: ID!, type: DamItemType!, folderId: ID, includeArchived: Boolean, filter: DamItemFilterInput): Float!
   damFilesList(offset: Int = 0, limit: Int = 20, sortColumnName: String, sortDirection: SortDirection = ASC, folderId: ID, includeArchived: Boolean = false, filter: FileFilterInput): PaginatedDamFiles!
   damFile(id: ID!): DamFile!
   damIsFilenameOccupied(filename: String!, folderId: String): Boolean!
@@ -464,14 +464,6 @@ input DamItemFilterInput {
 enum DamItemType {
   File
   Folder
-}
-
-input DamItemPositionInput {
-  sortColumnName: String
-  sortDirection: SortDirection = ASC
-  folderId: ID
-  includeArchived: Boolean
-  filter: DamItemFilterInput
 }
 
 input FileFilterInput {
