+++ packages/api/cms-api/schema.gql
@@ -219,7 +219,7 @@ type Query {
   pageTreeNodeList(scope: PageTreeNodeScopeInput!, category: String): [PageTreeNode!]!
   pageTreeNodeSlugAvailable(scope: PageTreeNodeScopeInput!, parentId: ID, slug: String!): SlugAvailability!
   damItemsList(offset: Int = 0, limit: Int = 20, sortColumnName: String, sortDirection: SortDirection = ASC, folderId: ID, includeArchived: Boolean, filter: DamItemFilterInput): PaginatedDamItems!
-  damItemListPosition(id: ID!, type: DamItemType!, args: DamItemPositionInput!): Float!
+  damItemListPosition(sortColumnName: String, sortDirection: SortDirection = ASC, id: ID!, type: DamItemType!, folderId: ID, includeArchived: Boolean, filter: DamItemFilterInput): Float!
   cronJobs: [CronJob!]!
   currentUser: CurrentUser!
 }
@@ -307,14 +307,6 @@ input DamItemFilterInput {
 enum DamItemType {
   File
   Folder
-}
-
-input DamItemPositionInput {
-  sortColumnName: String
-  sortDirection: SortDirection = ASC
-  folderId: ID
-  includeArchived: Boolean
-  filter: DamItemFilterInput
 }
 
 type Mutation {
