+++ packages/api/cms-api/src/common/sorting/sort.args.ts
@@ -1,22 +1,10 @@
-import { ArgsType, Field, InputType } from "@nestjs/graphql";
+import { ArgsType, Field } from "@nestjs/graphql";
 import { IsEnum, IsOptional, IsString } from "class-validator";
 
 import { SortDirection } from "./sort-direction.enum";
 
 @ArgsType()
 export class SortArgs {
-    @Field({ nullable: true })
-    @IsString()
-    @IsOptional()
-    sortColumnName?: string;
-
-    @Field(() => SortDirection, { defaultValue: SortDirection.ASC })
-    @IsEnum(SortDirection)
-    sortDirection?: SortDirection = SortDirection.ASC;
-}
-
-@InputType()
-export class SortInput {
     @Field({ nullable: true })
     @IsString()
     @IsOptional()
