+++ packages/api/cms-api/src/dam/files/dam-items.resolver.ts
@@ -1,22 +1,7 @@
-import { Args, createUnionType, Field, ID, Int, ObjectType, Query, registerEnumType, Resolver } from "@nestjs/graphql";
+import { Args, Field, Int, ObjectType, Query, Resolver } from "@nestjs/graphql";
 
 import { DamItemsService } from "./dam-items.service";
-import { DamItemPositionInput, DamItemsArgs } from "./dto/dam-items.args";
-import { File } from "./entities/file.entity";
-import { Folder } from "./entities/folder.entity";
-
-export const DamItem = createUnionType({
-    name: "DamItem",
-    types: () => [File, Folder] as const,
-});
-
-export enum DamItemType {
-    File = "file",
-    Folder = "folder",
-}
-registerEnumType(DamItemType, {
-    name: "DamItemType",
-});
+import { DamItem, DamItemPositionArgs, DamItemsArgs } from "./dto/dam-items.args";
 
 @ObjectType()
 class PaginatedDamItems {
@@ -43,11 +28,7 @@ export class DamItemsResolver {
     }
 
     @Query(() => Number)
-    async damItemListPosition(
-        @Args("id", { type: () => ID }) id: string,
-        @Args("type", { type: () => DamItemType }) type: DamItemType,
-        @Args("args", { type: () => DamItemPositionInput }) args: DamItemPositionInput,
-    ): Promise<number> {
-        return this.damItemsService.getDamItemPosition(id, type, args);
+    async damItemListPosition(@Args() args: DamItemPositionArgs): Promise<number> {
+        return this.damItemsService.getDamItemPosition(args);
     }
 }
