+++ packages/api/cms-api/src/dam/files/dam-items.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from "@nestjs/common";
 
-import { DamItem, DamItemType } from "./dam-items.resolver";
-import { DamItemPositionInput, DamItemsArgs } from "./dto/dam-items.args";
+import { DamItem, DamItemPositionArgs, DamItemsArgs, DamItemType } from "./dto/dam-items.args";
 import { FilesService } from "./files.service";
 import { FoldersService } from "./folders.service";
 
@@ -50,7 +49,7 @@ export class DamItemsService {
         return [response, foldersTotalCount + filesTotalCount];
     }
 
-    async getDamItemPosition(id: string, type: DamItemType, args: DamItemPositionInput): Promise<number> {
+    async getDamItemPosition({ type, id, ...args }: DamItemPositionArgs): Promise<number> {
         if (type === DamItemType.Folder) {
             const folderPosition = await this.foldersService.getFolderPosition(id, {
                 parentId: args.folderId,
