+++ packages/api/cms-api/src/dam/files/dto/file.args.ts
@@ -3,7 +3,7 @@ import { Type } from "class-transformer";
 import { IsBoolean, IsOptional, IsString, IsUUID, ValidateNested } from "class-validator";
 
 import { OffsetBasedPaginationArgs } from "../../../common/pagination/offset-based.args";
-import { SortArgs, SortInput } from "../../../common/sorting/sort.args";
+import { SortArgs } from "../../../common/sorting/sort.args";
 
 @InputType()
 export class FileFilterInput {
@@ -37,8 +37,7 @@ export class FileArgs extends IntersectionType(OffsetBasedPaginationArgs, SortAr
     filter?: FileFilterInput;
 }
 
-@InputType()
-export class DamFileListPositionInput extends SortInput {
+export class DamFileListPositionArgs extends SortArgs {
     @Field(() => ID, { nullable: true })
     @IsOptional()
     @IsUUID()
