+++ packages/api/cms-api/src/dam/files/dto/folder.args.ts
@@ -3,7 +3,7 @@ import { Type } from "class-transformer";
 import { IsBoolean, IsOptional, IsString, IsUUID, ValidateNested } from "class-validator";
 
 import { OffsetBasedPaginationArgs } from "../../../common/pagination/offset-based.args";
-import { SortArgs, SortInput } from "../../../common/sorting/sort.args";
+import { SortArgs } from "../../../common/sorting/sort.args";
 
 @InputType()
 export class FolderFilterInput {
@@ -44,8 +44,7 @@ export class FolderByNameAndParentIdArgs {
     parentId?: string;
 }
 
-@InputType()
-export class DamFolderListPositionInput extends SortInput {
+export class DamFolderListPositionArgs extends SortArgs {
     @Field(() => ID, { nullable: true })
     @IsOptional()
     @IsUUID()
